All buttons are actionable and do something. This is essentially GUI v1.0.
Replaced the stock ActionBar with a more professional Toolbar that will be consistent across all the APIs.
Added a splash screen that'll be presented whilst the app starts - this makes the loading process feel more seamless.
Insured that the toolbar was complIant with Google's reccomendations (Material Design).
Built all of the Activities and .xml files to coincide with the buttons; templates are availabe to be extended.

The work is commented with headings but feel free to let me know if this is insufficient information.