Book Appointments
Can now book appointments successfully (provided user does not have a
password). Beginning to work on date arithmetic (currently only shows
appointments on 2015-01-01 and 2015-02-02). After these are finished the
activity will be provided with a starting date, and be able to return
appointments for that date, and appointments for dates within a week of
that date.